Bug lorsque la poutre n'est pas orienté sur l'axe des X est résolu

+++ src/structure/include/structure/structure.hpp
@@ -14,7 +14,7 @@ struct Settings {
   std::string Mesh_file_path;
   std::string Solve_type;
   double Tolerance;
-  double N_step;
+  int N_step;
   double Damping; 
 };
